requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Data analysis
pandas>=2.1.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Data storage
openpyxl>=3.1.0  # Excel support

# Dashboard
streamlit>=1.30.0
bcrypt>=4.0.0

# Utilities
tqdm>=4.66.0  # Progress bars
